+++ packages/squiggle-lang/__tests__/SqProject/SqProject_tutorial_3_imports_test.ts
@@ -1,5 +1,5 @@
 import { SqProject } from "../../src/index.js";
-import { toStringResult } from "../../src/public/SqValue.js";
+import { toStringResult } from "../../src/public/SqValue/index.js";
 
 /**
  * Case: Imports.
