+++ packages/squiggle-lang/jest.config.js
@@ -4,6 +4,7 @@ const jestConfig = {
   testEnvironment: "node",
   moduleNameMapper: {
     "^(\\.{1,2}/.*)\\.js$": "$1",
+    "^@/(.*)\\.js$": "<rootDir>/src/$1",
   },
   transform: {
     "^.+\\.[jt]sx?$": [
