+++ packages/squiggle-lang/src/public/SqValue/SqDistribution/SqPointSet.ts
@@ -1,10 +1,12 @@
 import zipWith from "lodash/zipWith.js";
-import { SqPointSetDistribution } from "./SqDistribution.js";
-import { ContinuousShape } from "../PointSet/Continuous.js";
-import { DiscreteShape } from "../PointSet/Discrete.js";
-import { MixedShape } from "../PointSet/Mixed.js";
-import { AnyPointSet } from "../PointSet/PointSet.js";
-import { PointSetDist } from "../dist/PointSetDist.js";
+
+import { PointSetDist } from "@/dist/PointSetDist.js";
+import { ContinuousShape } from "@/PointSet/Continuous.js";
+import { DiscreteShape } from "@/PointSet/Discrete.js";
+import { MixedShape } from "@/PointSet/Mixed.js";
+import { AnyPointSet } from "@/PointSet/PointSet.js";
+
+import { SqPointSetDistribution } from "./index.js";
 
 enum Tag {
   Mixed = "Mixed",
