+++ packages/squiggle-lang/src/public/SqValue/SqDistribution/index.ts
@@ -1,13 +1,14 @@
-import { SampleSetDist } from "../dist/SampleSetDist/index.js";
-import { Env } from "../dist/env.js";
+import { SampleSetDist } from "@/dist/SampleSetDist/index.js";
+import { Env } from "@/dist/env.js";
+import * as Result from "@/utility/result.js";
+import { Ok, result } from "@/utility/result.js";
+import { BaseDist } from "@/dist/BaseDist.js";
+import { DistError } from "@/dist/DistError.js";
+import { SymbolicDist } from "@/dist/SymbolicDist.js";
+import { PointSetDist } from "@/dist/PointSetDist.js";
+
 import { SqDistributionError } from "./SqDistributionError.js";
 import { wrapPointSet } from "./SqPointSet.js";
-import * as Result from "../utility/result.js";
-import { BaseDist } from "../dist/BaseDist.js";
-import { DistError } from "../dist/DistError.js";
-import { SymbolicDist } from "../dist/SymbolicDist.js";
-import { PointSetDist } from "../dist/PointSetDist.js";
-import { Ok, result } from "../utility/result.js";
 
 export enum SqDistributionTag {
   PointSet = "PointSet",
