+++ packages/squiggle-lang/src/public/SqValue/SqPlot.ts
@@ -1,18 +1,19 @@
-import { BaseDist } from "../dist/BaseDist.js";
-import { SampleSetDist } from "../dist/SampleSetDist/index.js";
-import { Env } from "../dist/env.js";
-import * as Result from "../utility/result.js";
-import { Plot, vPlot } from "../value/index.js";
+import { BaseDist } from "@/dist/BaseDist.js";
+import { SampleSetDist } from "@/dist/SampleSetDist/index.js";
+import { Env } from "@/dist/env.js";
+import * as Result from "@/utility/result.js";
+import { Plot, vPlot } from "@/value/index.js";
+
+import { SqError, SqOtherError } from "../SqError.js";
+import { SqValuePath } from "../SqValuePath.js";
 import {
   SqDistribution,
   SqSampleSetDistribution,
   wrapDistribution,
-} from "./SqDistribution.js";
-import { SqError, SqOtherError } from "./SqError.js";
+} from "./SqDistribution/index.js";
 import { SqLambda } from "./SqLambda.js";
 import { SqScale, wrapScale } from "./SqScale.js";
-import { SqPlotValue } from "./SqValue.js";
-import { SqValuePath } from "./SqValuePath.js";
+import { SqPlotValue } from "./index.js";
 
 type LabeledSqDistribution = {
   name?: string;
